+++ src/libs/regexp/ipv4.mjs
@@ -1,6 +1,8 @@
-const PATTERN_COMPONENT = `(25[0-5]|2[0-4][0-9]|[0-1]?[0-9]{1,2})`;
-const PATTERN = `${PATTERN_COMPONENT}(\\.${PATTERN_COMPONENT}){3}`;
+import {DEC_OCTET} from "./rfc";
 
+const IPV4ADDRESS = `${DEC_OCTET}(\\.${DEC_OCTET}){3}`;
+
+const PATTERN = IPV4ADDRESS;
 const REGEXP = new RegExp(`^${PATTERN}$`);
 
 export {PATTERN, REGEXP};