+++ test/adjusters/string.test.mjs
@@ -241,6 +241,156 @@ function testMaxLength()
  */
 function testPattern()
 {
+	describe("ipv4", () =>
+	{
+		testPatternIpv4();
+	});
+	describe("ipv6", () =>
+	{
+		testPatternIpv6();
+	});
+	describe("uri", () =>
+	{
+		testPatternUri();
+	});
+	describe("others", () =>
+	{
+		testPatternOthers();
+	});
+}
+
+/**
+ * pattern; IPv4
+ * @returns {void}
+ */
+function testPatternIpv4()
+{
+	it("should be OK", () =>
+	{
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+			.adjust("0.0.0.0")).toEqual("0.0.0.0");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+			.adjust("192.168.0.1")).toEqual("192.168.0.1");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+			.adjust("255.255.255.255")).toEqual("255.255.255.255");
+	});
+	it("should cause error(s)", () =>
+	{
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+				.adjust("0.0.0.");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+				.adjust("0.0.0.0.");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+				.adjust("255.255.255.256");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV4)
+				.adjust("999.255.255.255");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+	});
+}
+
+/**
+ * pattern; IPv6
+ * @returns {void}
+ */
+function testPatternIpv6()
+{
+	it("should be OK", () =>
+	{
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+			.adjust("0000:0000:0000:0000:0000:0000:0000:0000")).toEqual("0000:0000:0000:0000:0000:0000:0000:0000");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+			.adjust("::1")).toEqual("::1");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+			.adjust("::")).toEqual("::");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+			.adjust("1::1")).toEqual("1::1");
+
+		// IPv4-mapped address
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+			.adjust("::ffff:192.0.2.1")).toEqual("::ffff:192.0.2.1");
+	});
+	it("should cause error(s)", () =>
+	{
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+				.adjust("0000");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+				.adjust("ffff:");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.IPV6)
+				.adjust("0000:0000:0000:0000:0000:0000:0000:0000:");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+	});
+}
+
+/**
+ * pattern; URI
+ * @returns {void}
+ */
+function testPatternUri()
+{
+	it("should be OK", () =>
+	{
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("https://example.com/path/to/resource?name=value#hash")).toEqual("https://example.com/path/to/resource?name=value#hash");
+
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("https://192.168.10.2/")).toEqual("https://192.168.10.2/");
+
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("https://[fe80::a1b3:125d:c1f8:4781]/")).toEqual("https://[fe80::a1b3:125d:c1f8:4781]/");
+
+		// https://tools.ietf.org/html/rfc3986#section-1.1.2
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("ftp://ftp.is.co.za/rfc/rfc1808.txt")).toEqual("ftp://ftp.is.co.za/rfc/rfc1808.txt");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("http://www.ietf.org/rfc/rfc2396.txt")).toEqual("http://www.ietf.org/rfc/rfc2396.txt");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("ldap://[2001:db8::7]/c=GB?objectClass?one")).toEqual("ldap://[2001:db8::7]/c=GB?objectClass?one");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("mailto:John.Doe@example.com")).toEqual("mailto:John.Doe@example.com");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("news:comp.infosystems.www.servers.unix")).toEqual("news:comp.infosystems.www.servers.unix");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("tel:+1-816-555-1212")).toEqual("tel:+1-816-555-1212");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("telnet://192.0.2.16:80/")).toEqual("telnet://192.0.2.16:80/");
+		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+			.adjust("urn:oasis:names:specification:docbook:dtd:xml:4.1.2")).toEqual("urn:oasis:names:specification:docbook:dtd:xml:4.1.2");
+	});
+	it("should cause error(s)", () =>
+	{
+		expect(() =>
+		{
+			adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
+				.adjust("https://例.com/");
+		}).toThrow(adjuster.CAUSE.PATTERN);
+	});
+}
+
+/**
+ * pattern; Others
+ * @returns {void}
+ */
+function testPatternOthers()
+{
 	it("should be OK", () =>
 	{
 		expect(adjuster.string().pattern(/^Go+gle$/)
@@ -254,9 +404,6 @@ function testPattern()
 
 		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.EMAIL)
 			.adjust("john.doe@example.com")).toEqual("john.doe@example.com");
-
-		expect(adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
-			.adjust("https://example.com/path/to/resource?name=value#hash")).toEqual("https://example.com/path/to/resource?name=value#hash");
 	});
 	it("should cause error(s)", () =>
 	{
@@ -276,12 +423,6 @@ function testPattern()
 		{
 			adjuster.string().pattern(adjuster.STRING.PATTERN.EMAIL)
 				.adjust("john..doe@example.com");
-		}).toThrow(adjuster.CAUSE.PATTERN);
-
-		expect(() =>
-		{
-			adjuster.string().pattern(adjuster.STRING.PATTERN.URI)
-				.adjust("https://例.com/");
 		}).toThrow(adjuster.CAUSE.PATTERN);
 	});
 }